+++ src/acceptance/org/smoothbuild/acceptance/lang/convert/ParameterDefaultArgumentConversionTest.java
@@ -4,7 +4,7 @@ import static org.smoothbuild.util.Strings.unlines;
 
 import org.smoothbuild.lang.base.type.TestedAssignment;
 
-public class ParameterDefaultValueConversionTest extends AbstractConversionTestCase {
+public class ParameterDefaultArgumentConversionTest extends AbstractConversionTestCase {
   @Override
   protected String createTestScript(TestedAssignment spec) {
     return unlines(