+++ src/main/java/org/smoothbuild/exec/compute/DefaultArgumentTask.java
@@ -14,12 +14,12 @@ import org.smoothbuild.lang.base.define.Location;
 import org.smoothbuild.lang.base.type.Type;
 import org.smoothbuild.util.Scope;
 
-public class DefaultValueTask extends StepTask {
+public class DefaultArgumentTask extends StepTask {
   private final int index;
   private final Scope<Task> scope;
   private final TaskCreator taskCreator;
 
-  public DefaultValueTask(Type type, String name, List<Task> dependencies, int index,
+  public DefaultArgumentTask(Type type, String name, List<Task> dependencies, int index,
       Location location, Scope<Task> scope, TaskCreator taskCreator) {
     super(BUILDER, type, "building:" + name, dependencies, location);
     this.index = index;
@@ -30,8 +30,7 @@ public class DefaultValueTask extends StepTask {
   @Override
   protected void onCompleted(Obj obj, Worker worker, Consumer<Obj> result) {
     String functionName = LambdaTuple.name(((Tuple) obj)).jValue();
-    Task task = taskCreator.namedFunctionParameterDefaultValueEagerTask(
-        scope, functionName, index);
+    Task task = taskCreator.defaultArgumentEagerTask(scope, functionName, index);
     task.compute(worker).addConsumer(result);
   }
 }